+++ Clex/AppRelease.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 enum AppRelease {
-    static let versionName = "1.9.3"
-    static let buildNumber = "193"
+    static let versionName = "1.9.5"
+    static let buildNumber = "195"
 }